src/test/kotlin/be/moac/aoc2021/Day03Test.kt
package be.moac.aoc2021

import org.assertj.core.api.Assertions.assertThat
import org.junit.jupiter.api.Test

internal class Day03Test {

    @Test
    internal fun `Part one`() {
        val input = listOf(
            "00100",
            "11110",
            "10110",
            "10111",
            "10101",
            "01111",
            "00111",
            "11100",
            "10000",
            "11001",
            "00010",
            "01010"
        )                         

        val result = Day03 partOne input

        assertThat(result).isEqualTo(198L)
    }

    @Test
    internal fun `part two`() {
        val input = listOf(
            "00100",
            "11110",
            "10110",
            "10111",
            "10101",
            "01111",
            "00111",
            "11100",
            "10000",
            "11001",
            "00010",
            "01010"
        )

        val result = Day03 partTwo input

        assertThat(result).isEqualTo(230)
    }

}
